requirements.txt
pyethereum
ethereum-serpent
requests